% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{BoxCox}
\alias{BoxCox}
\title{Box-Cox power transformation of a vector, adjusted to retain mean and SD.}
\usage{
BoxCox(x, lamda = c(-5, 5), standardise = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{lamda}{Either a single power parameter for the transformation
or a vector of length 2 giving the range of lamda values to search
to find an optimised transformation parameter.}

\item{standardise}{logical.  If TRUE, the returned value has mean=0 and SD=1.
Otherwise (the default) the original mean and SD are retained.}
}
\description{
Perform a power transformation on a numeric vector, retaining (by default)
its mean and standard deviation.
If two lamda values are given, this range is searched to find a power
transformation parameter that minimises \code{\link{skewness}}.
If \code{x} contains non-positive values, the vector is first shifted
into the positive Reals before the transformation.
}
\examples{
x = rexp(50); xnew = BoxCox(x)
plot(x, xnew, ylab=paste0('BoxCox(x) used lamda=',signif(attr(xnew,'lamda'),4)))
qqnorm(BoxCox(x))
}
