% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carryforward.R
\name{carryforward}
\alias{carryforward}
\alias{carryback}
\title{Carry forward/back imputation of a vector}
\usage{
carryforward(v)

carryback(v)
}
\arguments{
\item{v}{A numeric or character vector, or a factor, containing some missing values.}
}
\description{
Fill in missing values by carrying forward (or back) the known values in a vector or factor.
}
\section{Functions}{
\itemize{
\item \code{carryback}: The reverse
}}

\examples{
carryforward(c(NA,10,NA,23,19,NA,NA,NA,0,NA))
carryback(c(NA,10,NA,23,19,NA,NA,NA,0,NA))
}
