% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniques.R
\name{duplicates}
\alias{duplicates}
\title{Return values that are duplicates}
\usage{
duplicates(x, ...)
}
\arguments{
\item{x}{a vector.}

\item{...}{further parameters for \code{\link{duplicated}},
i.e. \code{incomparables=}.}
}
\description{
Given a vector, return the set of values that appear more than once.
}
\examples{
duplicates(c('cat','dog','dog'))
}
