% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{fuzzyround}
\alias{fuzzyround}
\title{Fuzzy (random) rounding.}
\usage{
fuzzyround(x)
}
\arguments{
\item{x}{a numeric vector or matrix.}
}
\description{
Values are rounded at random, with probabilities determined by
how close they are to the integers either side.
This may be useful for interesting colour dithering in images.
}
\examples{
x = seq(0,20,length.out=200); plot(x, fuzzyround(x))
image(fuzzyround(toeplitz(seq(5,1,length.out=200))), col=heat.colours())
}
