% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{gel}
\alias{gel}
\title{Transparent shades of grey or other colours}
\usage{
gel(alpha, red = 0, green = 0, blue = 0)
}
\arguments{
\item{alpha}{semi-transparency parameter (see \code{\link{rgb}})}

\item{red, green, blue}{values between 0 and 1 for colour components}
}
\description{
Inspired by gels used in theatre lighting.
}
\examples{
plot(rnorm(1000),rnorm(1000),pch=16,cex=3,col=gel(0.1))
}
