% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{heat.colours}
\alias{heat.colours}
\title{Some more realistic heat colours}
\usage{
heat.colours(n = 20, alpha = 1)
}
\arguments{
\item{n}{the number of colors to be in the palette.}

\item{alpha}{the alpha transparency.}
}
\description{
More realistic (or at least nicer) heat colours,
for use in e.g. \code{\link{image}} instead of \code{\link{heat.colors}}.
}
\details{
These colours go from almost black to white, via red and yellow.
Suggestions for improvement are most welcome!
}
\examples{
{
  layout(rbind(1:2))
  image(toeplitz(20:1), col=heat.colours(), main='heat.colours')
  image(toeplitz(20:1), col=heat.colors(20), main='heat.colors')
}
}
