% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.ascii.R
\name{is.ascii}
\alias{is.ascii}
\title{Identify non-ASCII elements in a character vector.}
\usage{
is.ascii(str)
}
\arguments{
\item{str}{A character vector.}
}
\description{
Given a character vector, returns a logical vector of the same length
that identifies which elements contain non-ASCII (>127) characters.
}
\examples{
print(str <- c('cat', 'caf\\u00e9'))
is.ascii(str)
}
