% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniques.R
\name{is.sorted}
\alias{is.sorted}
\title{Test whether a vector is sorted}
\usage{
is.sorted(x, ...)
}
\arguments{
\item{x}{a vector.}

\item{...}{further parameters for \code{\link{is.unsorted}}.}
}
\description{
Test whether a vector is sorted
}
\examples{
is.sorted(LETTERS)
}
