% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis of a vector}
\usage{
kurtosis(x, excess = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{excess}{logical; compute excess rather than absolute kurtosis?}

\item{na.rm}{logical; should missing values be removed?}
}
\description{
Computes the excess or absolute kurtosis of a vector.
By default, it returns the excess kurtosis, so that (for example)
the Normal distribution returns values near zero.
}
\examples{
kurtosis(rnorm(1000))
kurtosis(rnorm(1000), excess=FALSE)
kurtosis(runif(1000))
kurtosis(rcauchy(1000))
}
\seealso{
\link{skewness}
}
