% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like_unlike_functions.R
\name{\%like\%}
\alias{\%like\%}
\alias{\%Like\%}
\alias{\%unlike\%}
\alias{\%Unlike\%}
\title{Binary operators for regexp string comparisons.}
\usage{
x \%like\% y

x \%Like\% y

x \%unlike\% y

x \%Unlike\% y
}
\arguments{
\item{x}{character vector to search}

\item{y}{regular expression}
}
\description{
Inspired by the "like" operator in SQL,
these binary operators are a convenient way of using \code{\link{grepl}}.
}
\section{Functions}{
\itemize{
\item \code{\%like\%}: find matching values, ignoring case.

\item \code{\%Like\%}: case-sensitive matching.

\item \code{\%unlike\%}: finds non-matching values, ignoring case.

\item \code{\%Unlike\%}: finds non-matching values and is case-sensitive.
}}

\examples{
lets = c(letters, LETTERS)
lets[lets \%like\% '[aeiou]']
lets[lets \%Like\% '[aeiou]']
lets[lets \%unlike\% '[aeiou]']
lets[lets \%Unlike\% '[aeiou]']
}
\seealso{
\link{grepvi}
}
