% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{new.timing}
\alias{new.timing}
\title{Timing class for recording computation times}
\usage{
\method{new}{timing}(processes = "process1")
}
\arguments{
\item{processes}{character vector for names of processes to follow.}
}
\description{
The 'timing' class records the computation time of R calculations.
There are three methods: new, update and print.
}
\details{
Create an object of class 'timing', a list containing:
   now:     the last known value of sum(proc.time()[1:2])
   times:   named vector of processes to keep track of
   calls:   number of calls to proc.time within update.timing
}
