% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newpassword.R
\name{newpassword}
\alias{newpassword}
\title{Create a random password.}
\usage{
newpassword(Letters = 8, Numbers = 5, Symbols = 0)
}
\arguments{
\item{Letters}{Number of characters from \link{letters}/LETTERS to include.}

\item{Numbers}{Number of digits from 0-9 to include.}

\item{Symbols}{Number of symbol characters to include.}
}
\value{
A randomly generaged password.
}
\description{
Generates a random password containing the desired number of letters,
numeric digits and ASCII symbols.
}
\examples{
# Defaults; easy to double-click, copy and paste.
newpassword()
# A longer and more complicated password.
newpassword(9,7,5)
# Multiple new passwords.
replicate(10, newpassword())
}
