% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniques.R
\name{nuniques}
\alias{nuniques}
\title{Number of unique values}
\usage{
nuniques(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{...}{further parameters for \code{unique()}.}
}
\description{
Number of unique values, including NA as one of the values,
like \code{\link{unique}} does.
}
\examples{
nuniques(letters)
}
