% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{pastel}
\alias{pastel}
\title{Pastel colours}
\usage{
pastel(n, ...)
}
\arguments{
\item{n}{the number of colours to produce.}
}
\description{
Similar to \code{\link{rainbow}} but producing more muted colours,
derived using \code{\link{hcl}}.
}
\examples{
barplot(table(sample(LETTERS,1000,repl=TRUE)),col=pastel(26))
}
