% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plots}
\alias{plots}
\title{An informative array plots on a numeric vector}
\source{
Thanks to Paul Murrell for the original idea.
}
\usage{
plots(x, na.rm = !is.ts(x), title = NULL)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{logical; should NA values be removed?}

\item{title}{main title for the array of plots (derived from name of \code{x})}
}
\description{
Produces an array of interesting plots
to illustrate the values in the numeric vector \code{x}.
Includes several summary statistics.
}
\details{
The following plots are produced in a 2x3 array:

(1) Box Plot, with summary stats for number of observations,
missing/valid values, minimum, maximum, quartiles

(2) Histogram and Density, including overlayed Normal density

(3) Auto-correlation correlogram

(4) Normal Q-Q plot, with stats for mean, SD, \link{skewness} and
excess \link{kurtosis}

(5) Association of values at lag=1

(6) Run plot, with \link{loess} curve.
}
\examples{
# Requires a reasonably large plotting window:
plots(rnorm(1000))
plots(cumsum(rnorm(200)))

}
