% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{print.timing}
\alias{print.timing}
\title{Print method for class timing.}
\usage{
\method{print}{timing}(timing, ...)
}
\arguments{
\item{timing}{a 'timing' object}

\item{...}{additional arguments for print.default().}
}
\description{
Print method for class timing.
}
