% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{skewness}
\alias{skewness}
\title{Skewness of a vector}
\usage{
skewness(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{logical; should missing values be removed?}
}
\description{
Return the skewness of the distribution for a numeric vector.
}
\examples{
skewness(rnorm(1000))
skewness(rexp(1000))
}
\seealso{
\link{BoxCox}, \link{kurtosis}
}
