% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim leading and trailing white space}
\usage{
trim(txt)
}
\arguments{
\item{txt}{Character object}
}
\description{
Removes white space from the start and end of each element
in a character vector (or other similar character object, e.g. matrix).
}
\examples{
trim(c('    abc','xyz  ','\\t\\tindented.  '))
}
