% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniques.R
\name{uniques}
\alias{uniques}
\title{Sorted unique values}
\usage{
uniques(x, incomparables = FALSE, decreasing = FALSE, na.last = TRUE)
}
\arguments{
\item{x}{a vector.}

\item{incomparables}{as for \code{\link{unique}}.}

\item{decreasing, na.last}{as for \code{\link{sort}}.}
}
\description{
Returns the set of unique values, sorted.
}
\examples{
uniques(c(1, 1, NA, 1, -1))
}
