% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{update.timing}
\alias{update.timing}
\title{An update method for a timing object}
\usage{
\method{update}{timing}(timing, process = 1, ...)
}
\arguments{
\item{timing}{a 'timing' object.}

\item{process}{length 1; which process within timing to be updated (numeric index or name)}

\item{...}{not used.}
}
\description{
Assumes times[process] already exists (checking would take time).
Make sure length(process)==1.
Note that process can be a character(1) from names(timing$times).
}
