% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_functions.R
\name{Write.csv}
\alias{Write.csv}
\title{Write CSV file with different defaults}
\usage{
Write.csv(x, file, ...)
}
\arguments{
\item{x}{Matrix or data.frame}

\item{file}{File name, perhaps without .CSV extention. Defaults to name of first argument.}
}
\description{
Write a CSV file using \code{write.csv} (see \code{\link[utils]{write.table}}
with different defaults.
Specifically, NA values are empty, row names are excluded and the *.CSV file extention is added if not found in the file name.
The file name can default to a name derived from the name of the first argument.
}
\examples{
\dontrun{
Write.csv(iris)
}

}
