% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeropad.R
\name{zeropad}
\alias{zeropad}
\title{Format numbers in constant width with leading zeros}
\usage{
zeropad(x, digits = 0, width = NULL)
}
\arguments{
\item{x}{Vector of numbers.}

\item{digits}{Number of decimal places (defaults to 0).}

\item{width}{Width of character values returned.}
}
\description{
Formats numbers for alphabetic sorting by padding with zeros
to constant string width.
If width is not supplied, a default value is calculated
based on the values in \code{x}.
}
\examples{
paste0('ID', zeropad(1:20))
}
