% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_functions.R
\name{NoYes}
\alias{NoYes}
\title{Convert a logical vector to a factor}
\usage{
NoYes(bool, ifFALSE = "No", ifTRUE = "Yes")
}
\arguments{
\item{bool}{A logical vector, or a vector that can be coerced to logical.}

\item{ifFALSE}{factor level for FALSE values, defaults to "No".}

\item{ifTRUE}{factor level for TRUE values, defaults to "Yes".}
}
\description{
Converts a logical vector to a factor conatining two levels.
This is useful in statistical models such as logistic regression,
both for the dependent variable and the explanatory variables.
}
\examples{
table(NoYes(rnorm(100) > 0.5))
}
