% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_functions.R
\name{NoYes}
\alias{NoYes}
\title{Convert a logical vector to a two-level factor}
\usage{
NoYes(bool, levels = c("No", "Yes"))
}
\arguments{
\item{bool}{A logical vector, or a vector that can be coerced to logical.}

\item{levels}{Two factor levels for FALSE and TRUE values, respectively.}
}
\value{
A factor with two levels.
}
\description{
Converts a logical vector to a factor conatining two levels.
This is useful in statistical models such as logistic regression,
both for the dependent variable and the explanatory variables.
}
\examples{
table(NoYes(LETTERS \%in\% c('A','E','I','O','U')))
table(NoYes(runif(100) > 0.5, c('Tails', 'Heads')))
}
