% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_functions.R
\name{NoYes}
\alias{NoYes}
\title{Convert a logical vector to a factor}
\usage{
NoYes(bool)
}
\arguments{
\item{bool}{A logical vector, or a vector that can be coerced to logical.}
}
\value{
A factor with levels combining the unique levels of the input factors.
}
\description{
Convert a logical vector to a factor
}
